Refactoring: component ModalSearch
