+++ src/components/ModalSearch/types/ResponsData.ts
@@ -1,0 +1,1 @@
+type ResponsData<T> = { data: T[]; resLenght: number };
